Bluetooth data plotting and command parsing
Test sketch to collect faster data from the LSM6DS3 IMU (208Hz) and broadcast this over bluetooth via an attached HC-05. This data can be plotted in the Arduino IDE once the COM port is changed to the Bluetooth one. This will be useful to see how effective various PID controller settings are.

Another test sketch to parse commands received via the Bluetooth serial interface. This will be useful to dynamically set PID constants without having to upload new code to the Arduino.

Another test sketch showing how to receive input from the RC transmitter, which will eventually be used to control the speed and direction of the robot.