Improved balancing OffroadBalancingControl
Increased KI and KD for better stability, but the robot still oscillates
a bit. Also implemented deadband for speed to try to stop it drifting to
one side. The deadband keeps the speed at zero below a threshold. I
think this is responsible for the oscillation because the robot falls
a bit while the motors are stationary.