Improved balance control
Changed deadband handling logic to scale the speed so that it is always
higher than the configured deadband. Retuned the PID constants
accordingly. The robot can balance, but vibrates when close to
stationary. This is most likely due to backlash in the motor gearbox.
Maybe further tuning of the PID constants can reduce this.