Adjusted PID constants and encoder speed filtering
Tuned the PID constants to be less aggressive, but the result was pretty
much the same. The robot still vibrates due to backlash when stationary.

Wrote a test program to test the filtering of the encoder signal. The
delta value from the encoder is read at 250Hz and run through a 10Hz low
pass filter. The result is much less noisy than the raw integer-based
encoder value.