Added speed control using RC transmitter
Changed the balance auto-correction from a simple accumulator to a PID
controller. Used RC receiver to alter the speed setpoint so the robot
can move forwards or backwards.

The speed PID controller needs a lot more tuning as the movement is
very jerky, as was the case in the Barbie Segway. But it's a start at
least.