Changed speed SpeedFilterTest
Changed the speed filter from 10Hz cutoff to 4Hz cutoff and saw a
reduction in the noise level. This will hopefully allow KP and KI to be
increased on the speed PID controller.

Also made the steering linear depending on how far the RC steering wheel
is turned. The steering still feels a little sluggish, so maybe I'll
increase the multiplier.