Tuned speed PID controller for responsive driving
The robot starts moving straight away now rather than having to wait
several seconds. The speed fluctuates quite a lot, and there is
noticable oscillation while the robot drives.

The robot can go over small bumps and from hard floor to carpet. The
driving motion is smoother on carpet, at the expense of stability when
stationary. This hints that there is a tradeoff between the performance
of the balance PID controller and the speed PID controller. Reducing the
PID constants for the balance controller and increasing them for the
speed controller may be the path to better driving performance.