Tuned speed PID, increased max steering and speed
Reducing both kp and ki in the speed PID controller resulted in smoother
driving.

Increasing the deadband did not improve the performance of the robot at
all. Restesting on carpet showed similar speed harmonics as on a smooth
surface. Reducing kp and ki in the angle PID controller also did not
improve driving performance, and just increased the harmonics in the
speed.