Reduced both ki values
I guessed that the speed harmonics were due to too much integral action
from the controllers. For the angle controller, I decreased ki by 25%
and increased kp by 33%. This had minimal impact on the stability of the
robot when standing, but made the driving speed much smoother.

I also increased kp and reduced ki for the speed controller. The
acceleration is slower, but the driving motion is smoother.