Further tuning of angle PID controller
Increased KD in the angle PID controller to try to minimise the angle
overshoot when driving. This made a small difference to the smoothness
of the driving - see Data3, Data4 and Data7 worksheets in the
spreadsheet.

The speed controller does not seem to result in the robot reaching the
desired speed very often. Increasing KI in the speed controller should
help with this. However, I'm wondering whether the relationship between
angle and driving speed will be non-linear, so I want to try increasing
the maximum speed from 6 to 10 before tuning the speed controller
further.