Added low pass filter to accelerometer data
The pitch angle was noisy and I figured this might be due to inadequate
filtering of the accelerometer data. The LSM6DS3 apparently has a
built-in low pass filter, but I couldn't get it to enable so I coded one
myself. The result is a cleaner signal fed into the angle controller,
which will hopefully allow KD to be increased.

I also experimented with changing alpha in the complementary angle
filter from 0.996 to 0.96. The goal was to reduce the time taken to
correct gyro drift. However, the robot did not balance well (see
imu_data.xlsx). My theory is that the drift is reduced, and so the speed
controller changes the angle set point too quickly and the robot becomes
unstable. However, further investigation is needed to determine the root
cause.

I also increased the max steering PWM value from 70 to 100 to allow for
faster turning.