Tweaked speed PID controller
I tried increasing KD on the angle controller from 2 to 3, but this
resulted in more shaking when upright.

I then tried changing the speed controller KP from 0.3 to 0.25 and KI
from 0.15 to 0.18. This had minimal effect on the smoothness of the
driving, but is maybe _slightly_ better.

I also changed the startup code to again use the accelerometer data to
detect when the robot is upright to start balancing. The drift in the
complementary filter is too high to rely on the filtered value.