Increased top speed to 15
We want the robot to go fast because it's cooler. Increased the top
speed to 15 encoder counts per loop. The loop period is 4ms, there are
64 encoder counts per motor revolution, and the motor gearbox ratio is
30:1, so this translates to ~120rpm at the wheels (which is 1/3
of the max unloaded speed of the motors).

The steering correction still doesn't seem to work when driving
straight. I suspect this is because of spikes in the RC signal which
cause the sterring error to be reset (since it looks like the driver is
attempting to steer manually). We need some kind of notch filter to
reject any values which look like noise.