Improve remote control spike filtering
I noticed a couple of cases where the RC signal changed rapidly and the
robot was left perpetually turning in a circle. To avoid this, I changed
the spike detection to allow for only 2 spikes (where the old and new
reading differ by more than 400us) before using the signal as is. This
should eliminate the noise from the signal without the signal getting
locked on an incorrect value.