Allow steering from Raspberry Pi ball tracking
Added code to the Arduino OffroadBalancingRobot.ino to accept a
space-separated list of floating-point numbers. The first number is the
desired speed (in encoder counts per loop, min -5, max 5), and the second is
the speed (as a PWM value, min -40, max 40). The speed value is ignored for
now. Both the speed and steering serial commands are overridden by the RC
transmitter.

Added a new PID controller to ball-tracking.py that tries to keep the pan
angle of the camera at zero (i.e. the ball is right in front of the robot).
This currently uses a proportional component only to try to keep things
simple. The output of the PID controller is written to the serial channel to
the Arduino.