Use VideoCapture and perf_counter
Camera feed would stop about 20 seconds after the program started, i.e.
when the clock changed due to time sync. It's unclear why the imutils
VideoStream has this problem, but replacing it with a similar class that
uses OpenCV's VideoCapture to read the images from the camera solves the
issue. Also, used time.perf_counter instead of time.time so that the FPS
count is independent of the system clock.